requirements.txt
# Repo Rover - Python Dependencies

# Core Dependencies
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0

# Fetch.ai Agent Framework
uagents==0.12.0
cosmpy==0.9.2

# Google Gemini AI
google-generativeai==0.3.2

# Vectara RAG
vectara==0.1.2
# Alternative: Use requests for Vectara REST API

# ArXiv Paper Access
arxiv==2.1.0
PyPDF2==3.0.1

# GitHub Repository Handling
GitPython==3.1.40
pygithub==2.1.1

# Code Analysis
tree-sitter==0.20.4
tree-sitter-python==0.20.4

# Web Scraping (Fallback for Papers with Code)
beautifulsoup4==4.12.2
lxml==4.9.3

# Utility Libraries
rich==13.7.0  # Pretty terminal output
tqdm==4.66.1  # Progress bars
tenacity==8.2.3  # Retry logic

# Testing (Optional)
pytest==7.4.3
pytest-asyncio==0.21.1

# Jupyter Notebook Support (For Demo)
jupyter==1.0.0
ipywidgets==8.1.1
